# Add new internal `EditTree` data structure