--- a/internal/edittree/bitvector/README.md
+++ b/internal/edittree/bitvector/README.md
@@ -1,0 +1,15 @@
+# godropbox [![GoDoc](https://godoc.org/github.com/dropbox/godropbox?status.svg)](https://godoc.org/github.com/dropbox/godropbox) [![Actions Status](https://github.com/dropbox/godropbox/workflows/Test/badge.svg)](https://github.com/dropbox/godropbox/actions) [![Actions Status](https://github.com/dropbox/godropbox/workflows/Lint/badge.svg)](https://github.com/dropbox/godropbox/actions)
+
+Common libraries for writing go services/applications on Linux servers.
+
+### Requirements
+ * Go 1.13+
+ * Linux/x64
+
+### Installation
+``go get github.com/dropbox/godropbox``
+
+### Documentation
+
+See https://pkg.go.dev/github.com/dropbox/godropbox for modules documentation.
+