… internal/edittree/edittree.go
… internal/edittree/edittree_test.go
… vendor/github.com/kr/pretty/.gitignore
… vendor/github.com/kr/pretty/License
… vendor/github.com/kr/pretty/Readme
… vendor/github.com/kr/pretty/diff.go
… vendor/github.com/kr/pretty/formatter.go
… vendor/github.com/kr/pretty/pretty.go
… vendor/github.com/kr/pretty/zero.go
… vendor/github.com/kr/text/License
… vendor/github.com/kr/text/Readme
… vendor/github.com/kr/text/doc.go
… vendor/github.com/kr/text/indent.go
… vendor/github.com/kr/text/wrap.go
… vendor/gopkg.in/check.v1/.gitignore
… vendor/gopkg.in/check.v1/.travis.yml
… vendor/gopkg.in/check.v1/LICENSE
… vendor/gopkg.in/check.v1/README.md
… vendor/gopkg.in/check.v1/TODO
… vendor/gopkg.in/check.v1/benchmark.go
… vendor/gopkg.in/check.v1/check.go
… vendor/gopkg.in/check.v1/checkers.go
… vendor/gopkg.in/check.v1/helpers.go
… vendor/gopkg.in/check.v1/printer.go
… vendor/gopkg.in/check.v1/reporter.go
… vendor/gopkg.in/check.v1/run.go
… vendor/modules.txt