This PR separates out the new `EditTree` data structure that resulted from work on the recent [`json.patch` PR](https://github.com/open-policy-agent/opa/pull/5390). This PR should be merged before that one, as it's an upstream dependency.

-----

This PR adds a new `EditTree` data structure, which is built for applying in-place modifications to an `ast.Term`, and can render the final result of all of the edits efficiently.

The goal is to use this data structure to apply *all* of the patches in a JSON-Patch sequence rapidly, and then collapse all of the edits at the end with minimal wasted Term copying.
    
The `EditTree` data structure supports the following operations:

 - **Insert** :: Appends a new leaf node to the tree.
 - **Delete** :: Removes a leaf node from the tree.
 - **Unfold** :: Extends the tree down a given path, creating nodes as needed.
 - **Render** :: Generates the result term of tree node + all of its edits.
 - (Optional) **Exists** :: Fast boolean check for a path existing. Does not create new tree nodes.
 - (Optional) **Filter** :: Extracts a result term made from the provided paths (like `json.filter`).

It is designed to work across all of the core Rego data types, with special- case handling for scalars versus composite types.

Also! I have vendored in a [bit-vector library](https://github.com/dropbox/godropbox/tree/master/container/bitvector) to allow for ease-of-bookkeeping for in-place Array modifications. The indexing logic was especially gnarly in a few places around Array-specific areas.

-----

## Structure of the PR
This is a fairly beefy PR at around 5kLOC total. Below is the breakdown for what is probably most interesting to look at during code review:

 - Actual files for the data structure itself: 2000 LOC total
   - 1.1k LOC code (`internal/edittree/edittree.go`)
   - ~900 LOC tests (`internal/edittree/edittree_test.go`)
 - `json.filter` changes:
   - ~130 LOC benchmarks (`topdown/json_bench_test.go`)
   - ~15 LOC code (`topdown/json.go`)
 - Remainder of the PR is almost entirely just the vendored bit-vector library.

-----

## Remaining TODO Items

 - [x] Documentation around the new data structure
   - [x] Code-level documentation comments
   - [x] ~~Possibly a README.md in the directory, explaining how/why it works the way it does? (Could be useful for explaining the why behind certain design choices, rough algorithmic analysis, etc.)~~ EDIT: I've updated the comments within the `edittree.go` file's header area to be more informative about design choices that were made, and rough sketches of how each composite type is handled.
 - [x] Unit tests
   - Some unit tests already exist, but they are not exhaustive
   - Ideal would be full path coverage, but that's 16-20 cases per core operation at least.
     - Unit testing currently gets us >90% line coverage, which is good, but less thorough than path coverage would be.
 - [x] Benchmarks
   - [x] `json.filter` benchmark, which in this PR uses the `(*EditTree).Filter()` function.
     - Under-the-hood, this exercises around half of the `EditTree` APIs, particularly `Unfold`, `Exists`, and `Render`.
     - `Insert`/`Delete` will get benchmark exposure in the `json.patch` PR's benchmarks.