@philipaconrad 2023-01-06
I've realized that one of the stronger ways to deliver this PR is to have something it's being applied to.

In this case, I think the "end use application" is going to be `json.filter`, as local benchmark numbers look *very* promising. (~1.5-2x speedup for many use cases)

## `BenchmarkJSONFilterShallow` Results

`main`:
```
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -tags opa_wasm,slow -bench ^BenchmarkJSONFilterShallow$ github.com/open-policy-agent/opa/topdown

goos: linux
goarch: amd64
pkg: github.com/open-policy-agent/opa/topdown
cpu: Intel(R) Core(TM) i7-8650U CPU @ 1.90GHz
BenchmarkJSONFilterShallow/object-10-10-8         	   62730	     21069 ns/op	   10930 B/op	     246 allocs/op
BenchmarkJSONFilterShallow/object-100-10-8        	   40638	     25396 ns/op	   10930 B/op	     246 allocs/op
BenchmarkJSONFilterShallow/object-100-100-8       	   16195	     67784 ns/op	   34566 B/op	     889 allocs/op
BenchmarkJSONFilterShallow/object-1000-10-8       	   15892	     75183 ns/op	   10931 B/op	     246 allocs/op
BenchmarkJSONFilterShallow/object-1000-100-8      	    9648	    117596 ns/op	   34574 B/op	     889 allocs/op
BenchmarkJSONFilterShallow/object-1000-1000-8     	    2347	    482683 ns/op	  324936 B/op	    7223 allocs/op
BenchmarkJSONFilterShallow/object-10000-10-8      	    1929	    608117 ns/op	   10935 B/op	     246 allocs/op
BenchmarkJSONFilterShallow/object-10000-100-8     	    1788	    680947 ns/op	   34579 B/op	     889 allocs/op
BenchmarkJSONFilterShallow/object-10000-1000-8    	    1136	   1130732 ns/op	  325003 B/op	    7223 allocs/op
BenchmarkJSONFilterShallow/object-10000-10000-8   	     184	   5966788 ns/op	 3433534 B/op	   70420 allocs/op
PASS
ok  	github.com/open-policy-agent/opa/topdown	14.705s
```

This PR:
```
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -tags opa_wasm,slow -bench ^BenchmarkJSONFilterShallow$ github.com/open-policy-agent/opa/topdown

goos: linux
goarch: amd64
pkg: github.com/open-policy-agent/opa/topdown
cpu: Intel(R) Core(TM) i7-8650U CPU @ 1.90GHz
BenchmarkJSONFilterShallow/object-10-10-8         	   51192	     22465 ns/op	   11619 B/op	     246 allocs/op
BenchmarkJSONFilterShallow/object-100-10-8        	   63811	     19046 ns/op	   11619 B/op	     246 allocs/op
BenchmarkJSONFilterShallow/object-100-100-8       	   14431	     75659 ns/op	   37186 B/op	     794 allocs/op
BenchmarkJSONFilterShallow/object-1000-10-8       	   60570	     19589 ns/op	   11620 B/op	     246 allocs/op
BenchmarkJSONFilterShallow/object-1000-100-8      	   16884	     70641 ns/op	   37190 B/op	     794 allocs/op
BenchmarkJSONFilterShallow/object-1000-1000-8     	    1754	    708641 ns/op	  363686 B/op	    6202 allocs/op
BenchmarkJSONFilterShallow/object-10000-10-8      	   48553	     24383 ns/op	   11625 B/op	     246 allocs/op
BenchmarkJSONFilterShallow/object-10000-100-8     	   13825	     83671 ns/op	   37210 B/op	     795 allocs/op
BenchmarkJSONFilterShallow/object-10000-1000-8    	    1557	    803529 ns/op	  363765 B/op	    6202 allocs/op
BenchmarkJSONFilterShallow/object-10000-10000-8   	     134	   8913849 ns/op	 3660252 B/op	   60220 allocs/op
PASS
ok  	github.com/open-policy-agent/opa/topdown	18.121s
```

## `BenchmarkJSONFilterNested` Results

`main`:
```
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -tags opa_wasm,slow -bench ^BenchmarkJSONFilterNested$ github.com/open-policy-agent/opa/topdown

goos: linux
goarch: amd64
pkg: github.com/open-policy-agent/opa/topdown
cpu: Intel(R) Core(TM) i7-8650U CPU @ 1.90GHz
BenchmarkJSONFilterNested/object-10-10-8         	   40458	     33346 ns/op	   19440 B/op	     467 allocs/op
BenchmarkJSONFilterNested/object-100-10-8        	   37950	     31214 ns/op	   19440 B/op	     467 allocs/op
BenchmarkJSONFilterNested/object-100-100-8       	    5280	    206397 ns/op	  119446 B/op	    3090 allocs/op
BenchmarkJSONFilterNested/object-1000-10-8       	   36631	     32913 ns/op	   19440 B/op	     467 allocs/op
BenchmarkJSONFilterNested/object-1000-100-8      	    5774	    201381 ns/op	  119451 B/op	    3090 allocs/op
BenchmarkJSONFilterNested/object-1000-1000-8     	     526	   2123535 ns/op	 1173775 B/op	   29226 allocs/op
BenchmarkJSONFilterNested/object-10000-10-8      	   40995	     36315 ns/op	   19440 B/op	     467 allocs/op
BenchmarkJSONFilterNested/object-10000-100-8     	    4845	    227876 ns/op	  119448 B/op	    3090 allocs/op
BenchmarkJSONFilterNested/object-10000-1000-8    	     567	   1988200 ns/op	 1173819 B/op	   29226 allocs/op
BenchmarkJSONFilterNested/object-10000-10000-8   	      48	  24807449 ns/op	11914916 B/op	  290428 allocs/op
PASS
ok  	github.com/open-policy-agent/opa/topdown	14.814s
```

This PR:
```
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -tags opa_wasm,slow -bench ^BenchmarkJSONFilterNested$ github.com/open-policy-agent/opa/topdown

goos: linux
goarch: amd64
pkg: github.com/open-policy-agent/opa/topdown
cpu: Intel(R) Core(TM) i7-8650U CPU @ 1.90GHz
BenchmarkJSONFilterNested/object-10-10-8         	   51692	     25326 ns/op	   13144 B/op	     318 allocs/op
BenchmarkJSONFilterNested/object-100-10-8        	   47329	     24637 ns/op	   13144 B/op	     318 allocs/op
BenchmarkJSONFilterNested/object-100-100-8       	    9516	    108249 ns/op	   53309 B/op	    1582 allocs/op
BenchmarkJSONFilterNested/object-1000-10-8       	   50872	     30752 ns/op	   13144 B/op	     318 allocs/op
BenchmarkJSONFilterNested/object-1000-100-8      	   10641	    114042 ns/op	   53308 B/op	    1582 allocs/op
BenchmarkJSONFilterNested/object-1000-1000-8     	    1070	    963877 ns/op	  505353 B/op	   14186 allocs/op
BenchmarkJSONFilterNested/object-10000-10-8      	   47486	     29716 ns/op	   13144 B/op	     318 allocs/op
BenchmarkJSONFilterNested/object-10000-100-8     	    7287	    138072 ns/op	   53307 B/op	    1582 allocs/op
BenchmarkJSONFilterNested/object-10000-1000-8    	     895	   1267229 ns/op	  505349 B/op	   14186 allocs/op
BenchmarkJSONFilterNested/object-10000-10000-8   	     100	  11768706 ns/op	 5062002 B/op	  140200 allocs/op
PASS
ok  	github.com/open-policy-agent/opa/topdown	14.209s
```